requirements.txt
grpcio==1.16.1
h5py==2.9.0
imageio==2.6.0
jsonnet==0.14.0
mlflow==1.3.0
numpy==1.14.5
opencv-python==4.1.1.26
pillow==6.2.0
protobuf==3.9.2
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.1.0
tensorflow-gpu==1.13.2
tensorpack==0.9.8
urllib3==1.24.3
pandas
matplotlib